Move validation to base command for common resource
