+++ areas/applicationinsights/src/AzureMcp.ApplicationInsights/Commands/ListInsightsCommand.cs
@@ -59,31 +59,6 @@ public sealed class ListInsightsCommand(
         return options;
     }
 
-    public override ValidationResult Validate(CommandResult commandResult, CommandResponse? commandResponse = null)
-    {
-        var result = base.Validate(commandResult, commandResponse);
-
-        if (result.IsValid)
-        {
-            var resourceName = commandResult.GetValueForOption(ApplicationInsightsOptionDefinitions.ResourceName);
-            var resourceId = commandResult.GetValueForOption(ApplicationInsightsOptionDefinitions.ResourceId);
-
-            // Enforce that at least one of resourceName or resourceId is provided
-            if (string.IsNullOrWhiteSpace(resourceName) && string.IsNullOrWhiteSpace(resourceId))
-            {
-                result.IsValid = false;
-                result.ErrorMessage = "You must specify at least one of --resource-name or --resource-id for the Application Insights resource.";
-                if (commandResponse != null)
-                {
-                    commandResponse.Status = 400;
-                    commandResponse.Message = result.ErrorMessage;
-                }
-            }
-        }
-
-        return result;
-    }
-
     [McpServerTool(Destructive = false, ReadOnly = true, Title = CommandTitle)]
     public async override Task<CommandResponse> ExecuteAsync(CommandContext context, ParseResult parseResult)
     {
